Move Knight, Bishop, Rook and Queen classes to separate modules
