+++ src/pieces/main.js
@@ -15,192 +15,13 @@ limitations under the License.
 */
 
 
-const { Pawn } = require('./pawn');
 const { Piece, StepPiece, LinearPiece } = require('./base');
+const { Pawn } = require('./pawn');
+const { Knight } = require('./knight');
+const { Bishop } = require('./bishop');
+const { Rook } = require('./rook');
+const { Queen } = require('./queen');
 const { Relation } = require('../relations');
-
-
-class Knight extends StepPiece {
-    /*  Step points
-         ___ ___ ___ ___ ___
-        |   | B |   | C |   |
-       2|___|___|___|___|___|
-        | A |   |   |   | D |
-       1|___|___|___|___|___|
-        |   |   |Kni|   |   |
-       0|___|___|ght|___|___|
-        | H |   |   |   | E |
-      -1|___|___|___|___|___|
-        |   | G |   | F |   |
-      -2|___|___|___|___|___|
-          -2  -1   0   1   2
-    */
-    static stepPoints = [
-        {x: -2, y: 1},  // A
-        {x: -1, y: 2},  // B
-        {x: 1, y: 2},   // C
-        {x: 2, y: 1},   // D
-        {x: 2, y: -1},  // E
-        {x: 1, y: -2},  // F
-        {x: -1, y: -2}, // G
-        {x: -2, y: -1}, // H
-    ];
-
-    constructor(color, square, refresh=true) {
-        /*
-        Params:
-            color {string} (white or black);
-            square {Square} - where piece is placed;
-            refresh {boolean} - whether refresh board after piece placed or not.
-        */
-
-        super(color, square, Piece.KNIGHT, refresh);
-    }
-
-    getSquares() {
-        // Get knight squares by piece action.
-
-        this._refreshSquares();
-        this._getStepSquares(Knight.stepPoints);
-    }
-
-    getBind() {
-        // Bind knight.
-
-        this.getTotalImmobilize();
-    }
-}
-
-
-class Bishop extends LinearPiece {
-    /*  Directions
-         ___ ___ ___
-        | A |   | B |
-       1|___|___|___|
-        |   |Bis|   |
-       0|___|hop|___|
-        | D |   | C |
-      -1|___|___|___|
-          -1   0   1
-    */
-    static directions = [
-        {x: -1, y: 1},  // A (upleft)
-        {x: 1, y: 1},   // B (upright)
-        {x: 1, y: -1},  // C (downright)
-        {x: -1, y: -1}, // D (downleft)
-    ];
-
-    constructor(color, square, refresh=true) {
-        /*
-        Params:
-            color {string} (white or black);
-            square {Square} - where piece is placed;
-            refresh {boolean} - whether refresh board after piece placed or not.
-        */
-
-        super(color, square, Piece.BISHOP, refresh);
-    }
-
-    getSquares() {
-        // Get bishop squares by piece action.
-
-        this._refreshSquares();
-        this._getLinearSquares(Bishop.directions);
-    }
-}
-
-
-class Rook extends LinearPiece {
-    /*  Directions
-         ___ ___ ___
-        |   | A |   |
-       1|___|___|___|
-        | D |Ro | B |
-       0|___| ok|___|
-        |   | C |   |
-      -1|___|___|___|
-          -1   0   1
-    */
-    static directions = [
-        {x: 0, y: 1},  // A (up)
-        {x: 1, y: 0},  // B (right)
-        {x: 0, y: -1}, // C (down)
-        {x: -1, y: 0}, // D (left)
-    ];
-
-    constructor(color, square, refresh=true) {
-        /*
-        Params:
-            color {string} (white or black);
-            square {Square} - where piece is placed;
-            refresh {boolean} - whether refresh board after piece placed or not.
-        */
-
-        super(color, square, Piece.ROOK, refresh);
-    }
-
-    get castleRoad() {
-        if (this.hasOwnProperty('_castleRoad')) {
-            return this._castleRoad;
-        }
-        return null
-    }
-
-    getSquares() {
-        // Get rook squares by piece action.
-
-        this._refreshSquares();
-        this._getLinearSquares(Rook.directions);
-    }
-
-    setCastleRoad(castleRoad) {
-        /*
-        Params:
-            castleRoad {KingCastleRoad}.
-        */
-
-        this._castleRoad = castleRoad;
-    }
-
-    removeCastleRoad() {
-        this._castleRoad = null;
-    }
-}
-
-
-class Queen extends LinearPiece {
-    /*  Directions
-         ___ ___ ___
-        | A | B | C |
-       1|___|___|___|
-        | H |Que| D |
-       0|___| en|___|
-        | G | F | E |
-      -1|___|___|___|
-          -1   0   1
-
-        Actualy use Bishop and Rook directions.
-    */
-
-    constructor(color, square, refresh=true) {
-        /*
-        Params:
-            color {string} (white or black);
-            square {Square} - where piece is placed;
-            refresh {boolean} - whether refresh board after piece placed or not.
-        */
-
-        super(color, square, Piece.QUEEN, refresh);
-    }
-
-    getSquares() {
-        // Get queen squares by piece action.
-
-        this._refreshSquares();
-        this._getLinearSquares(Bishop.directions);
-        this._getLinearSquares(Rook.directions);
-    }
-}
 
 
 class KingCastleRoad {
